Revert 'Collection' header of #nav
This reverts commit c678674d7c7688a82b0b4afa3c0a979c3ca64a83.

Until I decide a better way to add this header, I'm going to revert it.
The problem is that, because of the way I had implemented it, this
header wasn't actually part of #nav, but instead replaced the Music All
Access banner. This conveniently meant that no new content had to be
added; the downside was that this header will not scroll with #nav, in
case the user resized the window and required that the #nav be
scrollable. Awkard.