Modify dark.css, CustomWebView to use/hold inverted sprite sheet (fixes #90, #120)
Previously, many things were made to match the dark style by using the
`-webkit-filter: invert(100%)` idiom in CSS, but this had a huge
performance impact since it was used for every thumbs up in every row,
on every player button, and so on.

To get rid of the performance hit, I instead download the sprite sheet
at the beginning of the program and invert the colors in memory. The
CSS is then updated to point to a special URL
(http://radiant-player-mac/sprites-inverted.png) at which point Radiant
Player returns the inverted image and the performance hit is now gone
as the inverted sprite sheet is now handled as a regular image.

In any forthcoming dark themes or derivatives of the built in theme, if
you need an inverted version of a particular sprite, set the
background-image to this special URL.