Hot-swap the bugged Web Components polyfill
The new Google Play Music design uses the Polymer framework, which
builds upon Web Components. To ensure compatibility, Google/Polymer
developed the `webcomponents.js` polyfill, but this apparently has a
bug in some versions of WebKit! The eventual outcome of the bug is that
the Flash audio player does not load properly.

Surprisingly, the bug was not present in my version of Safari, but it
was present in Cocoa's WebView (AppleWebKit), on the same system -- I
would have thought they shared the same web engine.

In particular, the bug was some strange difference in `__proto__` and
`Object.getPrototypeOf` in this particular version of WebKit on OS X.
In my version of Safari, these two are effectively the same when trying
to retrieve the prototype. However, for me and many users, the Cocoa
WebView returned an empty object for `__proto__` at one point instead
of returning the true prototype as `Object.getPrototypeOf` returns.

Until this gets investigated further and fixed upstream in
webcomponents.js, the temporary solution to this is to create my own
copy of webcomponents.js with the problematic usage of `__proto__`
replaced with `Object.getPrototypeOf`, and to provide my copy when
Google Play Music requests their copy.

TODO: Look into making HTML5 audio the default and obsoleting the Flash
Player.