Fix mini-player ghosting, and allow in full screen apps (fixes #79, fixes #309)
Such a simple fix for #79 that I can't even believe it took this long
-- in fact, I remember trying to investigate this issue and learning
that it was simply not possible without creating a UI agent
application. But apparently, adding
`NSWindowCollectionBehaviorFullScreenAuxiliary` does the trick.

The issue in #309 is that when switching spaces at any point after the
mini player has been opened and closed, a "ghost" of the mini player
will appear on the space being switched to, and disappears once the
space is fully switched. Changing to
`NSWindowCollectionBehaviorMoveToActiveSpace` means that the mini
player does not appear on all spaces, it simply just opens in the space
in which it is activated, which is almost the same use case.