macOS Sierra Fixes w/ MiniPlayer and CSS Updates
This fixes issues with macOS Sierra regarding cookie issues and a big
issue with the MiniPlayer which caused to crashed.

This also included fixed CSS Updates to all themes contained in Radiant
Player.

The following should be now fixed:

https://github.com/radiant-player/radiant-player-mac/issues/600

https://github.com/radiant-player/radiant-player-mac/issues/608

https://github.com/radiant-player/radiant-player-mac/issues/574

https://github.com/radiant-player/radiant-player-mac/issues/606

https://github.com/radiant-player/radiant-player-mac/issues/605

I also done a feature request from
https://github.com/radiant-player/radiant-player-mac/issues/607

But it is not included with this PR. I can add it to the PR if its
something needed or a decision among the Radiant Player Team.

The following PR has been tested on macOS Sierra GM, Yosemite, El
Capitan, and Mavericks!